=)
neue sachen

+++ Supernova/Supernova/Sub_Forms/Overview/FrmProjectOverview.cs
@@ -15,7 +15,7 @@ namespace Supernova.Sub_Forms.Overview
         DataLoad dl = new DataLoad();
         private int currentProjectID = 0;
         String currentProjectName;
-        private int currentPoints = 0;
+        private double currentPoints = 0;
         public FrmProjectOverview()
         {
             InitializeComponent();
@@ -50,7 +50,7 @@ namespace Supernova.Sub_Forms.Overview
             if (currentMouseOverRow != -1)
             {
                 currentProjectID = Convert.ToInt32(mainGrid[0, currentMouseOverRow].Value);
-                currentPoints = Convert.ToInt32(mainGrid[7, currentMouseOverRow].Value);
+                currentPoints = Convert.ToDouble(mainGrid[7, currentMouseOverRow].Value);
                 currentProjectName = mainGrid[1, currentMouseOverRow].Value.ToString();
 
             }
